requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
aiohttp=3.7.4.post0=pypi_0
async-timeout=3.0.1=pypi_0
attrs=20.3.0=pypi_0
beautifulsoup4=4.9.1=pypi_0
bs4=0.0.1=pypi_0
ca-certificates=2020.10.14=0
certifi=2020.6.20=pypi_0
chardet=3.0.4=pypi_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
discord=1.0.1=pypi_0
discord-py=1.7.1=pypi_0
googlesearch-python=2020.0.2=pypi_0
idna=2.10=pypi_0
imagehash=4.2.0=pypi_0
imageio=2.9.0=pypi_0
joblib=1.0.1=pypi_0
kiwisolver=1.3.1=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
lxml=4.6.3=pypi_0
matplotlib=3.4.1=pypi_0
multidict=5.1.0=pypi_0
ncurses=6.2=he6710b0_1
networkx=2.5.1=pypi_0
numpy=1.20.2=pypi_0
openssl=1.1.1k=h27cfd23_0
pandas=1.2.4=pypi_0
pillow=8.2.0=pypi_0
pip=21.0.1=py39h06a4308_0
profanity-check=1.0.6=pypi_0
pymongo=3.11.3=pypi_0
pyparsing=2.4.7=pypi_0
python=3.9.4=hdb3f193_0
python-dateutil=2.8.1=pypi_0
python-dotenv=0.17.1=pypi_0
pytz=2021.1=pypi_0
pywavelets=1.1.1=pypi_0
readline=8.1=h27cfd23_0
requests=2.24.0=pypi_0
scikit-image=0.18.1=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.6.3=pypi_0
setuptools=52.0.0=py39h06a4308_0
six=1.15.0=pypi_0
soupsieve=2.0.1=pypi_0
sqlite=3.35.4=hdfb4753_0
threadpoolctl=2.1.0=pypi_0
tifffile=2021.4.8=pypi_0
tk=8.6.10=hbc83047_0
tqdm=4.61.0=pypi_0
typing-extensions=3.7.4.3=pypi_0
tzdata=2020f=h52ac0ba_0
urllib3=1.25.9=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yarl=1.6.3=pypi_0
zlib=1.2.11=h7b6447c_3
